requirements.txt
aiofiles
aiosqlite
asynctempfile
discord.py